+++ packages/suspense/src/hooks/useCacheMutation.ts
@@ -44,7 +44,7 @@ export function useCacheMutation<Params extends Array<any>, Value>(
 
   const mutateSync = useCallback<MutateSync<Params, Value>>(
     (params: Params, newValue: Value) => {
-      const cacheKey = getKey(...params);
+      const cacheKey = getKey(params);
 
       if (mutationAbortControllerMap.has(cacheKey)) {
         const abortController = mutationAbortControllerMap.get(cacheKey)!;
@@ -69,7 +69,7 @@ export function useCacheMutation<Params extends Array<any>, Value>(
 
   const mutateAsync = useCallback<MutateAsync<Params, Value>>(
     async (params: Params, callback: MutationCallback<Value>) => {
-      const cacheKey = getKey(...params);
+      const cacheKey = getKey(params);
 
       if (mutationAbortControllerMap.has(cacheKey)) {
         const abortController = mutationAbortControllerMap.get(cacheKey)!;
@@ -98,7 +98,7 @@ export function useCacheMutation<Params extends Array<any>, Value>(
       mutationAbortControllerMap.set(cacheKey, abortController);
 
       startTransition(() => {
-        notifySubscribers(...params);
+        notifySubscribers(params);
         refresh(createPendingMutationRecordMap, pendingMutationRecordMap);
       });
 
@@ -131,7 +131,7 @@ export function useCacheMutation<Params extends Array<any>, Value>(
         }
 
         startTransition(() => {
-          notifySubscribers(...params);
+          notifySubscribers(params);
           refresh(createPendingMutationRecordMap, pendingMutationRecordMap);
         });
       } catch (error) {
@@ -150,7 +150,7 @@ export function useCacheMutation<Params extends Array<any>, Value>(
         recordMap.set(cacheKey, record);
 
         startTransition(() => {
-          notifySubscribers(...params);
+          notifySubscribers(params);
           refresh(createPendingMutationRecordMap, pendingMutationRecordMap);
         });
       } finally {