+++ packages/suspense/src/hooks/useIntervalCacheStatus.test.tsx
@@ -11,19 +11,43 @@ import {
   STATUS_REJECTED,
   STATUS_RESOLVED,
 } from "../constants";
-import { createCache } from "../cache/createCache";
-import { Cache, CacheLoadOptions, Deferred, Status } from "../types";
-import { useCacheStatus } from "./useCacheStatus";
 import { createDeferred } from "../utils/createDeferred";
+import { createIntervalCache } from "../cache/createIntervalCache";
+import {
+  Deferred,
+  IntervalCache,
+  IntervalCacheLoadOptions,
+  Status,
+} from "../types";
+import { useIntervalCacheStatus } from "./useIntervalCacheStatus";
 
-describe("useCacheStatus", () => {
-  let cache: Cache<[string], string>;
-  let fetch: jest.Mock<Promise<string> | string, [[string], CacheLoadOptions]>;
-  let getCacheKey: jest.Mock<string, [[string]]>;
+function createContiguousArray(start: number, end: number) {
+  return Array.from({ length: end - start + 1 }, (_, index) => start + index);
+}
+
+function getPointForValue(value: number) {
+  return value;
+}
+
+describe("useIntervalCacheStatus", () => {
+  let cache: IntervalCache<number, [id: string], number>;
+  let load: jest.Mock<
+    PromiseLike<number[]>,
+    [start: number, end: number, id: string, options: IntervalCacheLoadOptions]
+  >;
   let lastRenderedStatus: Status | undefined = undefined;
 
-  function Component({ string }: { string: string }): any {
-    lastRenderedStatus = useCacheStatus(cache, string);
+  function Component({
+    end,
+    start,
+    string,
+  }: {
+    end: number;
+    start: number;
+    string: string;
+  }): any {
+    lastRenderedStatus = useIntervalCacheStatus(cache, start, end, string);
+
     return lastRenderedStatus as any;
   }
 
@@ -31,24 +55,14 @@ describe("useCacheStatus", () => {
     // @ts-ignore
     global.IS_REACT_ACT_ENVIRONMENT = true;
 
-    fetch = jest.fn();
-    fetch.mockImplementation(async ([key]) => {
-      if (key.startsWith("async")) {
-        return Promise.resolve(key);
-      } else if (key.startsWith("error")) {
-        return Promise.reject(key);
-      } else {
-        return key;
-      }
+    load = jest.fn();
+    load.mockImplementation(async (start: number, end: number) => {
+      return createContiguousArray(start, end);
     });
 
-    getCacheKey = jest.fn();
-    getCacheKey.mockImplementation(([key]) => key.toString());
-
-    cache = createCache<[string], string>({
-      debugLabel: "cache",
-      getKey: getCacheKey,
-      load: fetch,
+    cache = createIntervalCache<number, [id: string], number>({
+      getPointForValue,
+      load,
     });
 
     lastRenderedStatus = undefined;
@@ -58,20 +72,21 @@ describe("useCacheStatus", () => {
     const container = document.createElement("div");
     const root = createRoot(container);
     act(() => {
-      root.render(<Component string="test" />);
+      root.render(<Component start={1} end={10} string="test" />);
     });
 
     expect(lastRenderedStatus).toBe(STATUS_NOT_FOUND);
   });
 
   it("should transition from pending to resolved", async () => {
-    const promise = cache.readAsync("async");
+    const promise = cache.readAsync(1, 5, "test");
 
     const container = document.createElement("div");
     const root = createRoot(container);
     act(() => {
-      root.render(<Component string="async" />);
+      root.render(<Component start={1} end={5} string="test" />);
     });
+
     expect(lastRenderedStatus).toBe(STATUS_PENDING);
 
     await act(async () => await promise);
@@ -80,65 +95,71 @@ describe("useCacheStatus", () => {
   });
 
   it("should transition from pending to rejected", async () => {
-    const promise = cache.readAsync("error");
+    const deferred = createDeferred<number[]>();
+    load.mockReturnValueOnce(deferred.promise);
+
+    const promise = cache.readAsync(1, 5, "test");
 
     const container = document.createElement("div");
     const root = createRoot(container);
     act(() => {
-      root.render(<Component string="error" />);
+      root.render(<Component start={1} end={5} string="test" />);
     });
+
     expect(lastRenderedStatus).toBe(STATUS_PENDING);
 
-    await act(async () => {
-      try {
-        await promise;
-      } catch (error) {}
-    });
+    await act(async () => deferred.reject(new Error("Expected")));
+    await expect(() => promise).rejects.toThrow("Expected");
 
     expect(lastRenderedStatus).toBe(STATUS_REJECTED);
   });
 
   it("should return resolved for keys that have already been loaded", async () => {
-    const promise = cache.readAsync("sync");
-    await promise;
+    await cache.readAsync(1, 5, "test");
 
     const container = document.createElement("div");
     const root = createRoot(container);
     act(() => {
-      root.render(<Component string="sync" />);
+      root.render(<Component start={1} end={5} string="test" />);
     });
+
     expect(lastRenderedStatus).toBe(STATUS_RESOLVED);
   });
 
   it("should return rejected for keys that have already failed", async () => {
-    try {
-      await cache.readAsync("error");
-    } catch (error) {}
+    const deferred = createDeferred<number[]>();
+    load.mockReturnValueOnce(deferred.promise);
+
+    const promise = cache.readAsync(1, 5, "error");
+
+    deferred.reject(new Error("Expected"));
+    await expect(() => promise).rejects.toThrow("Expected");
 
     const container = document.createElement("div");
     const root = createRoot(container);
     act(() => {
-      root.render(<Component string="error" />);
+      root.render(<Component start={1} end={5} string="error" />);
     });
+
     expect(lastRenderedStatus).toBe(STATUS_REJECTED);
   });
 
   it("should update in response to an aborted request", async () => {
     let abortSignal: AbortSignal | undefined;
-    let deferred: Deferred<string> | undefined;
-    fetch.mockImplementation(async (params, options) => {
+    let deferred: Deferred<number[]> | undefined;
+    load.mockImplementation(async (start, end, id, options) => {
       abortSignal = options.signal;
       deferred = createDeferred();
       return deferred.promise;
     });
 
-    cache.readAsync("async");
-    expect(cache.getStatus("async")).toBe(STATUS_PENDING);
+    cache.readAsync(1, 5, "async");
+    expect(cache.getStatus(1, 5, "async")).toBe(STATUS_PENDING);
 
     const container = document.createElement("div");
     const root = createRoot(container);
     act(() => {
-      root.render(<Component string="async" />);
+      root.render(<Component start={1} end={5} string="async" />);
     });
 
     expect(lastRenderedStatus).toBe(STATUS_PENDING);
@@ -146,10 +167,9 @@ describe("useCacheStatus", () => {
     act(() => {
       expect(cache.abort("async")).toBe(true);
     });
+
     expect(abortSignal?.aborted).toBe(true);
-
     await Promise.resolve();
-
     expect(lastRenderedStatus).toBe(STATUS_NOT_FOUND);
   });
 });