+++ packages/suspense/src/index.ts
@@ -6,6 +6,7 @@ export * from "./cache/createStreamingCache";
 export * from "./hooks/useCacheMutation";
 export * from "./hooks/useCacheStatus";
 export * from "./hooks/useImperativeCacheValue";
+export * from "./hooks/useIntervalCacheStatus";
 export * from "./hooks/useStreamingValue";
 export * from "./utils/assertRecordStatus";
 export * from "./utils/createDeferred";
@@ -13,7 +14,7 @@ export * from "./utils/createInfallibleCache";
 export * from "./utils/isPromiseLike";
 export * from "./utils/isRecordStatus";
 export * from "./utils/parallelize";
-export * from './utils/WeakRefMap'
+export * from "./utils/WeakRefMap";
 export * from "./utils/Record";
 export * from "./utils/RecordData";